requirements.txt
pytorch
numpy
plotly
pytorch